chore(store): Move helper functions to specific file
To avoid circular dependencies, it's best to move
helpers from the `index` file to a new file and
then export these functions to the index.
